ProjectDemo: fix most remaining warnings
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/CoreObject@9513 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Samples/ProjectDemo/OutlineController.m
@@ -271,7 +271,8 @@ static int i = 0;
 	assert(indexOfItemInParent != NSNotFound);
 	if (parent != nil && indexOfItemInParent > 0)
 	{
-		NSLog(@"Requesting object at %d in collection of %d", indexOfItemInParent - 1, (int)[[parent contents] count]);
+		NSLog(@"Requesting object at %d in collection of %d",
+			  (int)(indexOfItemInParent - 1), (int)[[parent contents] count]);
 		OutlineItem *newParent = [[parent contents] objectAtIndex: (indexOfItemInParent - 1)];
 		
 		if ([newParent isKindOfClass: [OutlineItem class]])
